+++ lib/collection.js
@@ -7,11 +7,6 @@ const ObjectID = require('mongodb-core').BSON.ObjectID;
 const MongoError = require('mongodb-core').MongoError;
 const toError = require('./utils').toError;
 const normalizeHintField = require('./utils').normalizeHintField;
-const handleCallback = require('./utils').handleCallback;
-const decorateCommand = require('./utils').decorateCommand;
-const decorateWithCollation = require('./utils').decorateWithCollation;
-const decorateWithReadConcern = require('./utils').decorateWithReadConcern;
-const formattedOrderClause = require('./utils').formattedOrderClause;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const CommandCursor = require('./command_cursor');
 const unordered = require('./bulk/unordered');
@@ -43,6 +38,7 @@ const DistinctOperation = require('./operations/distinct');
 const DropIndexOperation = require('./operations/drop_index');
 const DropIndexesOperation = require('./operations/drop_indexes');
 const EstimatedDocumentCountOperation = require('./operations/estimated_document_count');
+const FindOperation = require('./operations/find');
 const FindOneOperation = require('./operations/find_one');
 const FindOneAndDeleteOperation = require('./operations/find_one_and_delete');
 const FindOneAndReplaceOperation = require('./operations/find_one_and_replace');
@@ -89,8 +85,6 @@ const executeOperationV2 = require('./operations/execute_operation_v2');
  *   });
  * });
  */
-
-const mergeKeys = ['ignoreUndefined'];
 
 /**
  * Create a new Collection instance (INTERNAL TYPE, do not instantiate directly)
@@ -290,149 +284,9 @@ Collection.prototype.find = deprecateOptions(
       console.warn('Third parameter to `find()` must be a callback or undefined');
     }
 
-    let selector = query;
-    // figuring out arguments
-    if (typeof callback !== 'function') {
-      if (typeof options === 'function') {
-        callback = options;
-        options = undefined;
-      } else if (options == null) {
-        callback = typeof selector === 'function' ? selector : undefined;
-        selector = typeof selector === 'object' ? selector : undefined;
-      }
-    }
+    const findOperation = new FindOperation(this, query, options);
 
-    // Ensure selector is not null
-    selector = selector == null ? {} : selector;
-    // Validate correctness off the selector
-    const object = selector;
-    if (Buffer.isBuffer(object)) {
-      const object_size = object[0] | (object[1] << 8) | (object[2] << 16) | (object[3] << 24);
-      if (object_size !== object.length) {
-        const error = new Error(
-          'query selector raw message size does not match message header size [' +
-            object.length +
-            '] != [' +
-            object_size +
-            ']'
-        );
-        error.name = 'MongoError';
-        throw error;
-      }
-    }
-
-    // Check special case where we are using an objectId
-    if (selector != null && selector._bsontype === 'ObjectID') {
-      selector = { _id: selector };
-    }
-
-    if (!options) options = {};
-
-    let projection = options.projection || options.fields;
-
-    if (projection && !Buffer.isBuffer(projection) && Array.isArray(projection)) {
-      projection = projection.length
-        ? projection.reduce((result, field) => {
-            result[field] = 1;
-            return result;
-          }, {})
-        : { _id: 1 };
-    }
-
-    // Make a shallow copy of options
-    let newOptions = Object.assign({}, options);
-
-    // Make a shallow copy of the collection options
-    for (let key in this.s.options) {
-      if (mergeKeys.indexOf(key) !== -1) {
-        newOptions[key] = this.s.options[key];
-      }
-    }
-
-    // Unpack options
-    newOptions.skip = options.skip ? options.skip : 0;
-    newOptions.limit = options.limit ? options.limit : 0;
-    newOptions.raw = typeof options.raw === 'boolean' ? options.raw : this.s.raw;
-    newOptions.hint =
-      options.hint != null ? normalizeHintField(options.hint) : this.s.collectionHint;
-    newOptions.timeout = typeof options.timeout === 'undefined' ? undefined : options.timeout;
-    // // If we have overridden slaveOk otherwise use the default db setting
-    newOptions.slaveOk = options.slaveOk != null ? options.slaveOk : this.s.db.slaveOk;
-
-    // Add read preference if needed
-    newOptions.readPreference = resolveReadPreference(newOptions, {
-      db: this.s.db,
-      collection: this
-    });
-
-    // Set slave ok to true if read preference different from primary
-    if (
-      newOptions.readPreference != null &&
-      (newOptions.readPreference !== 'primary' || newOptions.readPreference.mode !== 'primary')
-    ) {
-      newOptions.slaveOk = true;
-    }
-
-    // Ensure the query is an object
-    if (selector != null && typeof selector !== 'object') {
-      throw MongoError.create({ message: 'query selector must be an object', driver: true });
-    }
-
-    // Build the find command
-    const findCommand = {
-      find: this.s.namespace,
-      limit: newOptions.limit,
-      skip: newOptions.skip,
-      query: selector
-    };
-
-    // Ensure we use the right await data option
-    if (typeof newOptions.awaitdata === 'boolean') {
-      newOptions.awaitData = newOptions.awaitdata;
-    }
-
-    // Translate to new command option noCursorTimeout
-    if (typeof newOptions.timeout === 'boolean') newOptions.noCursorTimeout = newOptions.timeout;
-
-    decorateCommand(findCommand, newOptions, ['session', 'collation']);
-
-    if (projection) findCommand.fields = projection;
-
-    // Add db object to the new options
-    newOptions.db = this.s.db;
-
-    // Add the promise library
-    newOptions.promiseLibrary = this.s.promiseLibrary;
-
-    // Set raw if available at collection level
-    if (newOptions.raw == null && typeof this.s.raw === 'boolean') newOptions.raw = this.s.raw;
-    // Set promoteLongs if available at collection level
-    if (newOptions.promoteLongs == null && typeof this.s.promoteLongs === 'boolean')
-      newOptions.promoteLongs = this.s.promoteLongs;
-    if (newOptions.promoteValues == null && typeof this.s.promoteValues === 'boolean')
-      newOptions.promoteValues = this.s.promoteValues;
-    if (newOptions.promoteBuffers == null && typeof this.s.promoteBuffers === 'boolean')
-      newOptions.promoteBuffers = this.s.promoteBuffers;
-
-    // Sort options
-    if (findCommand.sort) {
-      findCommand.sort = formattedOrderClause(findCommand.sort);
-    }
-
-    // Set the readConcern
-    decorateWithReadConcern(findCommand, this, options);
-
-    // Decorate find command with collation options
-    try {
-      decorateWithCollation(findCommand, this, options);
-    } catch (err) {
-      if (typeof callback === 'function') return callback(err, null);
-      throw err;
-    }
-
-    const cursor = this.s.topology.cursor(this.s.namespace, findCommand, newOptions);
-
-    return typeof callback === 'function' ? handleCallback(callback, null, cursor) : cursor;
+    return findOperation.execute(callback);
   }
 );
